core/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "core"
version = "0.1.0"
description = "Common Code"
authors = ["Jason Colburne <j.colburne@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "*"
uvicorn = "*"
gunicorn = "*"
toml = "*"
asyncpg = "*"
sqlmodel = "*"
python-multipart = "*"
python-jose = {extras = ["cryptography"], version = "*"}
pydantic = {extras = ["email"], version = "*"}
fastapi-utils = "*"
ipython = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"
black = "*"
pylint = "*"
mypy = "*"
asyncpg-stubs = "*"
pytest-asyncio = "*"
